requirements.txt
asttokens==2.4.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
comet-ml==3.35.1
comm==0.1.4
configobj==5.0.8
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dulwich==0.21.6
everett==3.1.0
executing==2.0.1
fonttools==4.44.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
jedi==0.19.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.10.6
scipy==1.11.3
seaborn==0.13.0
semantic-version==2.10.0
sentry-sdk==1.34.0
simplejson==3.19.2
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
websocket-client==1.3.3
widgetsnbextension==4.0.9
wrapt==1.15.0
wurlitzer==3.0.3
xgboost==2.0.1
